refactor(perry-ext-net): split tls + ip helpers into sibling modules (#1852)
The #1852 lifecycle additions pushed crates/perry-ext-net/src/lib.rs from
1991 to 2084 lines, over the 2000-line CI gate. Extract two cohesive,
low-coupling blocks into sibling modules (logic unchanged):

- src/tls.rs — rustls client config + do_tls_handshake (no shared net
  state; do_tls_handshake is pub(crate)).
- src/ip.rs — net.isIP/isIPv4/isIPv6 + auto-select-family accessors
  (pure helpers; uses crate::string_from_header_i64, now pub(crate)).

lib.rs is back to 1913 lines. Verified: ext-net checks + unit tests pass,
isIP and the full net lifecycle test still match Node.
