+++ crates/perry-ext-net/src/lib.rs
@@ -38,7 +38,6 @@ use perry_ffi::{
 use std::collections::HashMap;
 use std::io;
 use std::pin::Pin;
-use std::sync::Arc;
 use std::sync::Mutex;
 use std::task::{Context, Poll};
 
@@ -46,7 +45,15 @@ use tokio::io::{AsyncRead, AsyncReadExt, AsyncWrite, AsyncWriteExt, ReadBuf};
 use tokio::net::{TcpListener, TcpStream};
 use tokio::sync::{mpsc, oneshot};
 
-use tokio_rustls::{client::TlsStream, rustls, TlsConnector};
+use tokio_rustls::client::TlsStream;
+
+// #1852 — topical sub-modules split out to keep this file under the
+// 2000-line size gate. `tls` holds the rustls config + handshake; `ip`
+// holds the `net.isIP*` + auto-select-family helpers.
+mod ip;
+mod tls;
+
+use crate::tls::do_tls_handshake;
 
 // ─── Transport enum (plain or TLS, swappable at runtime) ─────────────────────
 
@@ -235,97 +242,7 @@ enum PendingNetEvent {
 
 // ─── Helpers ─────────────────────────────────────────────────────────────────
 
-/// Issue #811 — `net.isIP(s)` returns 0/4/6 (number).
-fn classify_ip(s: &str) -> i32 {
-    if is_ipv4_str(s) {
-        4
-    } else if is_ipv6_str(s) {
-        6
-    } else {
-        0
-    }
-}
-
-fn is_ipv4_str(s: &str) -> bool {
-    s.parse::<std::net::Ipv4Addr>().is_ok()
-}
-
-fn is_ipv6_str(s: &str) -> bool {
-    // Node's `net.isIPv6` rejects brackets and zone-id (`%`) suffixes —
-    // those forms aren't bare addresses.
-    if s.contains('[') || s.contains(']') || s.contains('%') {
-        return false;
-    }
-    s.parse::<std::net::Ipv6Addr>().is_ok()
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn js_net_is_ip(s_ptr: i64) -> f64 {
-    let kind = match string_from_header_i64(s_ptr) {
-        Some(s) => classify_ip(&s),
-        None => 0,
-    };
-    f64::from_bits(JsValue::from_number(kind as f64).0)
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn js_net_is_ipv4(s_ptr: i64) -> f64 {
-    let is = match string_from_header_i64(s_ptr) {
-        Some(s) => is_ipv4_str(&s),
-        None => false,
-    };
-    f64::from_bits(JsValue::from_bool(is).0)
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn js_net_is_ipv6(s_ptr: i64) -> f64 {
-    let is = match string_from_header_i64(s_ptr) {
-        Some(s) => is_ipv6_str(&s),
-        None => false,
-    };
-    f64::from_bits(JsValue::from_bool(is).0)
-}
-
-// Happy-Eyeballs (auto-select-family) defaults. Process-wide globals
-// that `getDefault*` reads and `setDefault*` writes.
-static AUTO_SELECT_FAMILY: std::sync::atomic::AtomicBool = std::sync::atomic::AtomicBool::new(true);
-// Node's current default is 500ms (raised from 250 in v20.x); pin to
-// 500 so byte-for-byte parity holds against `node --experimental-strip-types`.
-static AUTO_SELECT_FAMILY_ATTEMPT_TIMEOUT_MS: std::sync::atomic::AtomicI32 =
-    std::sync::atomic::AtomicI32::new(500);
-
-#[no_mangle]
-pub unsafe extern "C" fn js_net_get_default_auto_select_family() -> f64 {
-    let v = AUTO_SELECT_FAMILY.load(std::sync::atomic::Ordering::Relaxed);
-    f64::from_bits(JsValue::from_bool(v).0)
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn js_net_set_default_auto_select_family(val_f64: f64) -> f64 {
-    let val = JsValue(val_f64.to_bits()).to_bool();
-    AUTO_SELECT_FAMILY.store(val, std::sync::atomic::Ordering::Relaxed);
-    f64::from_bits(JsValue::UNDEFINED.0)
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn js_net_get_default_auto_select_family_attempt_timeout() -> f64 {
-    let v = AUTO_SELECT_FAMILY_ATTEMPT_TIMEOUT_MS.load(std::sync::atomic::Ordering::Relaxed);
-    f64::from_bits(JsValue::from_number(v as f64).0)
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn js_net_set_default_auto_select_family_attempt_timeout(ms_f64: f64) -> f64 {
-    let n = JsValue(ms_f64.to_bits()).to_number();
-    let ms = if n.is_finite() && n >= 0.0 {
-        n as i32
-    } else {
-        0
-    };
-    AUTO_SELECT_FAMILY_ATTEMPT_TIMEOUT_MS.store(ms, std::sync::atomic::Ordering::Relaxed);
-    f64::from_bits(JsValue::UNDEFINED.0)
-}
-
-unsafe fn string_from_header_i64(ptr: i64) -> Option<String> {
+pub(crate) unsafe fn string_from_header_i64(ptr: i64) -> Option<String> {
     let p = ptr as usize;
     if p < 0x1000 {
         return None;
@@ -547,94 +464,6 @@ fn mark_closed(id: i64) {
     if let Some(s) = statics::sockets().lock().unwrap().get_mut(&id) {
         s.is_open = false;
     }
-}
-
-// ─── rustls config ───────────────────────────────────────────────────────────
-
-fn build_tls_connector(verify: bool) -> Result<TlsConnector, String> {
-    if !verify {
-        return build_tls_connector_insecure();
-    }
-    let mut root_store = rustls::RootCertStore::empty();
-    let native = rustls_native_certs::load_native_certs();
-    for cert in native.certs {
-        let _ = root_store.add(cert);
-    }
-    let config = rustls::ClientConfig::builder()
-        .with_root_certificates(root_store)
-        .with_no_client_auth();
-    Ok(TlsConnector::from(Arc::new(config)))
-}
-
-fn build_tls_connector_insecure() -> Result<TlsConnector, String> {
-    use rustls::client::danger::{HandshakeSignatureValid, ServerCertVerified, ServerCertVerifier};
-    use rustls::pki_types::{CertificateDer, ServerName, UnixTime};
-    use rustls::{DigitallySignedStruct, SignatureScheme};
-
-    #[derive(Debug)]
-    struct NoVerify;
-
-    impl ServerCertVerifier for NoVerify {
-        fn verify_server_cert(
-            &self,
-            _end_entity: &CertificateDer<'_>,
-            _intermediates: &[CertificateDer<'_>],
-            _server_name: &ServerName<'_>,
-            _ocsp: &[u8],
-            _now: UnixTime,
-        ) -> Result<ServerCertVerified, rustls::Error> {
-            Ok(ServerCertVerified::assertion())
-        }
-        fn verify_tls12_signature(
-            &self,
-            _message: &[u8],
-            _cert: &CertificateDer<'_>,
-            _dss: &DigitallySignedStruct,
-        ) -> Result<HandshakeSignatureValid, rustls::Error> {
-            Ok(HandshakeSignatureValid::assertion())
-        }
-        fn verify_tls13_signature(
-            &self,
-            _message: &[u8],
-            _cert: &CertificateDer<'_>,
-            _dss: &DigitallySignedStruct,
-        ) -> Result<HandshakeSignatureValid, rustls::Error> {
-            Ok(HandshakeSignatureValid::assertion())
-        }
-        fn supported_verify_schemes(&self) -> Vec<SignatureScheme> {
-            vec![
-                SignatureScheme::RSA_PKCS1_SHA256,
-                SignatureScheme::RSA_PKCS1_SHA384,
-                SignatureScheme::RSA_PKCS1_SHA512,
-                SignatureScheme::ECDSA_NISTP256_SHA256,
-                SignatureScheme::ECDSA_NISTP384_SHA384,
-                SignatureScheme::RSA_PSS_SHA256,
-                SignatureScheme::RSA_PSS_SHA384,
-                SignatureScheme::RSA_PSS_SHA512,
-                SignatureScheme::ED25519,
-            ]
-        }
-    }
-
-    let config = rustls::ClientConfig::builder()
-        .dangerous()
-        .with_custom_certificate_verifier(Arc::new(NoVerify))
-        .with_no_client_auth();
-    Ok(TlsConnector::from(Arc::new(config)))
-}
-
-async fn do_tls_handshake(
-    tcp: TcpStream,
-    servername: &str,
-    verify: bool,
-) -> Result<TlsStream<TcpStream>, String> {
-    let connector = build_tls_connector(verify)?;
-    let server_name = rustls::pki_types::ServerName::try_from(servername.to_string())
-        .map_err(|e| format!("invalid servername '{}': {}", servername, e))?;
-    connector
-        .connect(server_name, tcp)
-        .await
-        .map_err(|e| format!("tls handshake: {}", e))
 }
 
 // ─── Spawning helper ─────────────────────────────────────────────────────────
